# api/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Taxi Platform API",
    "version": "1.0.0",
    "description": "Backend API contract for authentication, rides, admin operations."
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Local development server"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/rides": {
      "post": {
        "tags": [
          "rides"
        ],
        "summary": "Create a new ride",
        "operationId": "createRide",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RideCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ride created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ride"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/rides/calculate-price": {
      "post": {
        "tags": [
          "rides"
        ],
        "summary": "Estimate ride price and duration",
        "operationId": "calculateRidePrice",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PriceCalculationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Price estimation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceCalculationResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/rides/active": {
      "get": {
        "tags": [
          "rides"
        ],
        "summary": "Get current active ride",
        "operationId": "getActiveRide",
        "responses": {
          "200": {
            "description": "Active ride or null",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActiveRideResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/rides/history": {
      "get": {
        "tags": [
          "rides"
        ],
        "summary": "Get ride history",
        "operationId": "getRideHistory",
        "responses": {
          "200": {
            "description": "Ride history",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Ride"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/rides/available": {
      "get": {
        "tags": [
          "rides"
        ],
        "summary": "Get available rides for driver",
        "operationId": "getAvailableRides",
        "responses": {
          "200": {
            "description": "Available rides",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RideAvailable"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/rides/{rideId}/cancel": {
      "post": {
        "tags": [
          "rides"
        ],
        "summary": "Cancel ride",
        "operationId": "cancelRide",
        "parameters": [
          {
            "$ref": "#/components/parameters/RideId"
          }
        ],
        "responses": {
          "200": {
            "description": "Ride cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/rides/{rideId}": {
      "get": {
        "tags": [
          "rides"
        ],
        "summary": "Get ride by id",
        "operationId": "getRideById",
        "parameters": [
          {
            "$ref": "#/components/parameters/RideId"
          }
        ],
        "responses": {
          "200": {
            "description": "Ride details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ride"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/rides/{rideId}/status": {
      "patch": {
        "tags": [
          "rides"
        ],
        "summary": "Update ride status",
        "operationId": "updateRideStatus",
        "parameters": [
          {
            "$ref": "#/components/parameters/RideId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RideStatusUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ride updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ride"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/rides/{rideId}/accept": {
      "post": {
        "tags": [
          "rides"
        ],
        "summary": "Accept ride by driver",
        "operationId": "acceptRide",
        "parameters": [
          {
            "$ref": "#/components/parameters/RideId"
          }
        ],
        "responses": {
          "200": {
            "description": "Ride accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ride"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/rides/{rideId}/complete": {
      "post": {
        "tags": [
          "rides"
        ],
        "summary": "Complete ride by driver",
        "operationId": "completeRide",
        "parameters": [
          {
            "$ref": "#/components/parameters/RideId"
          }
        ],
        "responses": {
          "200": {
            "description": "Ride completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ride"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "RideId": {
        "name": "rideId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "RideStatus": {
        "type": "string",
        "enum": [
          "CREATED",
          "OFFERED",
          "ACCEPTED",
          "IN_PROGRESS",
          "COMPLETED",
          "CANCELLED",
          "EXPIRED"
        ]
      },
...
      "MessageResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "Location": {
        "type": "object",
        "required": [
          "lat",
          "lng"
        ],
        "properties": {
          "lat": {
            "type": "number",
            "format": "double"
          },
          "lng": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "CarType": {
        "type": "string",
        "enum": [
          "STANDTART",
          "BUSINESS",
          "XL"
        ]
      },
      "Coordinates": {
        "type": "object",
        "required": [
          "lat",
          "lng"
        ],
        "properties": {
          "lat": {
            "type": "number",
            "format": "double"
          },
          "lng": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "RideCreateRequest": {
        "type": "object",
        "required": [
          "pickupLocation",
          "dropoffLocation",
          "carType"
        ],
        "properties": {
          "pickupLocation": {
            "$ref": "#/components/schemas/Location"
          },
          "dropoffLocation": {
            "$ref": "#/components/schemas/Location"
          },
          "carType": {
            "$ref": "#/components/schemas/CarType"
          }
        }
      },
      "RideStatusUpdateRequest": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "$ref": "#/components/schemas/RideStatus"
          }
        }
      },
      "PriceCalculationRequest": {
        "type": "object",
        "required": [
          "pickupLocation",
          "dropoffLocation"
        ],
        "properties": {
          "pickupLocation": {
            "$ref": "#/components/schemas/Coordinates"
          },
          "dropoffLocation": {
            "$ref": "#/components/schemas/Coordinates"
          }
        }
      },
      "PriceCalculationResponse": {
        "type": "object",
        "required": [
          "distance",
          "duration",
          "price",
          "currency"
        ],
        "properties": {
          "distance": {
            "type": "number",
            "format": "double"
          },
          "duration": {
            "type": "integer"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "currency": {
            "type": "string"
          }
        }
      },
      "Ride": {
        "type": "object",
        "required": [
          "id",
          "passengerId",
          "passengerName",
          "pickupLocation",
          "dropoffLocation",
          "status",
          "price",
          "distance",
          "estimatedTime",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "passengerId": {
            "type": "string",
            "format": "uuid"
          },
          "passengerName": {
            "type": "string"
          },
          "driverId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "driverName": {
            "type": "string",
            "nullable": true
          },
          "pickupLocation": {
            "$ref": "#/components/schemas/Location"
          },
          "dropoffLocation": {
            "$ref": "#/components/schemas/Location"
          },
          "status": {
            "$ref": "#/components/schemas/RideStatus"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "distance": {
            "type": "number",
            "format": "double"
          },
          "estimatedTime": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "acceptedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "ActiveRideResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Ride"
          }
        ],
        "nullable": true
      },
      "RideAvailable": {
        "type": "object",
        "required": [
          "id",
          "passengerName",
          "pickupLocation",
          "dropoffLocation",
          "status",
          "price",
          "distance"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "passengerName": {
            "type": "string"
          },
          "pickupLocation": {
            "$ref": "#/components/schemas/Location"
          },
          "dropoffLocation": {
            "$ref": "#/components/schemas/Location"
          },
          "status": {
            "$ref": "#/components/schemas/RideStatus"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "distance": {
            "type": "number",
            "format": "double"
          },
          "estimatedTime": {
            "type": "integer",
            "nullable": true
          }
        }
      },
...
      "ValidationError": {
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
    }
  }
}